refactor: use store inside the tag, rename components and split them up a bit
