+++ packages/api-reference/src/components/Content/Tag/index.ts
@@ -1,3 +1,2 @@
-export { default as Tag } from './Tag.vue'
-export { default as TagAccordion } from './TagAccordion.vue'
+export { default as TagSection } from './TagSection.vue'
 export { default as TagList } from './TagList.vue'
